refactor(lib): move svg-sprite
