+++ app.js
@@ -28,7 +28,7 @@ const passport = require('passport')
 const controllers = require('./app/controllers')
 const requestHandlers = require('./lib/request_handlers')
 const initCloudinary = require('./app/lib/cloudinary')
-const compileSVGSprites = require('./lib/svg-sprite')
+const compileSVGSprites = require('./app/lib/svg-sprite')
 const appURL = require('./app/lib/url')
 const apiRoutes = require('./app/api_routes')
 const serviceRoutes = require('./app/service_routes')
